+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- GitHub actions is now working.
+
 ## [0.1.1] - 2024-02-02
 
 ### Fixed