+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- GitHub actions is now working.
+- GitHub actions are now working.
 
 ## [0.1.1] - 2024-02-02