+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - GitHub actions are now working.
 
+## [0.1.2] - 2024-02-03
+
+### Fixed
+
+- Fix bootstrap script parameter block.
+
 ## [0.1.1] - 2024-02-02
 
 ### Fixed
@@ -16,7 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Status badges updated.
 - Update integration tests for the bootstrap script.
 - Fix missing commands in bootstrap script.
-- Fix bootstrap script parameter block.
 
 ### Changed